requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
boto3==1.34.34
requests==2.31.0
python-multipart==0.0.9 